requirements.txt
python-telegram-bot==20.0
chatterbot==1.0.5
chatterbot_corpus==1.2.0
textblob==0.15.3